feat: add instagram rekap builder util
